Wily 13.42 tarball

+++ include/libg.h
@@ -91,12 +91,12 @@ struct	Fontchar
 
 struct	Subfont
 {
-	short		minrow;	/* first character row in font (for X subfonts) */
-	short		mincol;	/* first character col in font (for X subfonts) */
-	short		minchar; /* first char code in subfont */
-	short		maxchar; /* last char code in subfont */
-	short		width;	/* number of chars in row */
-	short		n;	/* number of chars in font */
+	int		minrow;	/* first character row in font (for X subfonts) */
+	int		mincol;	/* first character col in font (for X subfonts) */
+	int		minchar; /* first char code in subfont */
+	int		maxchar; /* last char code in subfont */
+	int		width;	/* number of chars in row */
+	int		n;	/* number of chars in font */
 	unsigned char	height;	/* height of bitmap */
 	char		ascent;	/* top of bitmap to baseline */
 	Fontchar	*info;	/* n+1 character descriptors */
